Now allows to simulate a range camera
Improvement: a range camera can be included as a new robot sensor with
any relative pose respect to the robot (but always with a null roll
angle). Its characteristics must be set in the code (they cannot be read
from config file).

The structure of the main class "CRobotSimulApp" has been slightly
modified, at least respect to the laser scanner and the range camera:
every sensor that can be simulated is represented by a new class (see
"CSimRangeCamera" or "CSimLaser"), and instances of these classes are
included in the main class. This gives rise to a tidier scheme and it's
also easier to expand/generalize if more kinds of sensors want to be
included in the future. Infrared and sonar has not been adapted yet.